# containers/swagger/schema/jamro-oversort-store-1.0.0-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/jamro/oversort-store/1.0.0"
  } ],
  "info" : {
    "description" : "REST API of OverSort Store",
    "version" : "1.0.0",
    "title" : "OverSort Store API"
  },
  "paths" : {
    "/sorted/{sortId}" : {
      "get" : {
        "summary" : "Get past sorting results",
        "operationId" : "history",
        "description" : "Each sort operation is stored and can be retrived later. This call returns past sort result basing on its sortId\n",
        "parameters" : [ {
          "in" : "path",
          "name" : "sortId",
          "description" : "reference ID of sort operation",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "sorted array",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SortList"
                },
                "example" : {
                  "sortId" : "bf04ec3e-1db2-4a43-9f6a-ce01ccf293dd",
                  "input" : [ "omega", "alpha", "beta" ],
                  "output" : [ "alpha", "beta", "omega" ]
                }
              }
            }
          },
          "400" : {
            "description" : "sortId parameter is missing"
          },
          "404" : {
            "description" : "Sort list was not found"
          },
          "500" : {
            "description" : "Internal error"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "SortList" : {
        "type" : "object",
        "properties" : {
          "sortId" : {
            "type" : "string",
            "description" : "Unique ID of sort operation"
          },
          "input" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "description" : "List of items to be sort"
          },
          "output" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "description" : "Sorted list"
          }
        }
      }
    }
  }
}